try to fix mmap and ummap

+++ os/src/syscall/mod.rs
@@ -105,7 +105,14 @@ pub fn syscall(syscall_id: usize, args: [usize; 6]) -> isize {
         SYSCALL_MUNMAP => 0,
         SYSCALL_CLONE => sys_fork(args[0], args[1], args[2], args[3], args[4]),
         SYSCALL_EXECVE => sys_execve(args[0] as *const u8, args[1] as *const usize),
-        SYSCALL_MMAP => sys_mmap(args[0], args[1], args[2], args[3], args[4], args[5]),
+        SYSCALL_MMAP => sys_mmap(
+            args[0], 
+            args[1], 
+            args[2] as u32, 
+            args[3] as u32, 
+            args[4], 
+            args[5]
+        ),
         SYSCALL_WAITPID => sys_waitpid(args[0] as isize, args[1] as *mut i32),
         SYSCALL_THREAD_CREATE => sys_thread_create(args[0], args[1]),
         SYSCALL_GETTID => sys_gettid(),
